refactor: 추후 분기 로직을 위해 커스텀 훅으로 분리 진행
